pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Jonathas Rodrigues <jonathasrc96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dynaconf = "^3.1.11"
PyGithub = "^1.55"
SQLAlchemy = "^1.4.41"

[tool.poetry.dev-dependencies]
# pytest = "^5.2"
pytest = "^6.0"
ipython = "^8.5.0"
ipdb = "^0.13.9"
isort = "^5.10.1"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
